backend/requirements.txt
Flask==2.3.2
moviepy==1.0.3
gTTS==2.3.2
deep-translator==1.11.4
supabase==2.4.5
openai-whisper==20231117
torch==2.1.0
numpy==1.26.4
ffmpeg-python==0.2.0
